change formatting

+++ 1/fahrenheit.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-  int fahr, celsius;
+  double fahr, celsius;
   int lower, upper, step;
   
   lower = 0;       /* lower limit of temperature table */
@@ -11,7 +11,7 @@ int main() {
   fahr = lower;
   while (fahr <= upper) {
     celsius = 5 * (fahr - 32) / 9;
-    printf("%d\t%d\n", fahr, celsius);
+    printf("%3.1f\t%6.1f\n", fahr, celsius);
     fahr = fahr + step;
   }
 }
